Split streams replication into connetion form and streams table page objects
